clean code fixes, PerformanceBrowse component, PerformanceByTheaters update, birthdate added to payload in auth.service
