+++ backend/src/middlewares/auth.middleware.js
@@ -24,7 +24,7 @@ export const isAuthenticated = (req, res, next) => {
 };
 
 export const isAdmin = (req, res, next) => {
-  if (req.user && req.user.role === "Admin") {
+  if (req.user && req.user.role === "admin") {
     next();
   } else {
     next(new HttpError("Admin role required", 403));
