+++ backend/src/services/auth.service.js
@@ -41,6 +41,7 @@ const login = async ({ email, password }) => {
     lastName: user.lastName,
     firstName: user.firstName,
     role: user.role,
+    birthDate: user.birthDate,
   };
 
   const token = jwt.sign(payload, JWT_SECRET);
