+++ frontend/src/components/NewPerformanceForm.jsx
@@ -6,7 +6,7 @@ import { toast } from 'react-toastify';
 import DefaultButton from './misc/DefaultButton';
 import performanceValidationSchema from '../schema/performanceValidationSchema';
 import getCreators from '../services/creators.service';
-import createPerformance from '../services/performance.service';
+import createPerformance from '../services/performances.service';
 import getTheaters from '../services/theaters.service';
 
 export default function NewPerformanceForm({ lecture }) {
