+++ frontend/src/components/performances/PerformancesByTheaters.jsx
@@ -29,18 +29,19 @@ export default function PerformancesByTheaters({ performances }) {
 
   // Csoportosítás színházak szerint
   const performancesByTheater = performances.reduce((groups, perf) => {
-    const theaterId = perf.theaterId || 'Ismeretlen színház';
-    if (!groups[theaterId]) {
-      groups[theaterId] = [];
+    const theaterName = perf.theater?.name || 'Ismeretlen színház'; // Színház neve
+    if (!groups[theaterName]) {
+      groups[theaterName] = [];
     }
-    groups[theaterId].push(perf);
+    groups[theaterName].push(perf);
     return groups;
   }, {});
 
   // Véletlenszerű színház kiválasztása
-  const theaterIds = Object.keys(performancesByTheater);
-  const randomTheaterId = theaterIds[Math.floor(Math.random() * theaterIds.length)];
-  const theaterPerformances = performancesByTheater[randomTheaterId];
+  const theaterNames = Object.keys(performancesByTheater);
+  const randomTheaterName =
+    theaterNames[Math.floor(Math.random() * theaterNames.length)];
+  const theaterPerformances = performancesByTheater[randomTheaterName];
 
   const scroll = (direction, containerRef) => {
     if (containerRef.current) {
@@ -76,8 +77,7 @@ export default function PerformancesByTheaters({ performances }) {
     <div className="w-full my-12 space-y-12">
       <section className="relative mb-12">
         <h2 className="text-2xl font-bold mb-5 text-c-text">
-          Random Színház ID:{' '}
-          {randomTheaterId === 'Ismeretlen színház' ? randomTheaterId : `#${randomTheaterId}`}
+          Random Színház: {randomTheaterName}
         </h2>
         <div className="relative">
           <div className="flex items-center justify-between">
