+++ frontend/src/pages/BrowsingPage.jsx
@@ -1,50 +1,44 @@
-import { useEffect, useRef, useState } from 'react';
-import { useSearchParams } from 'react-router-dom';
+import { useContext, useEffect, useRef, useState } from 'react';
 
 // Components
+import AuthContext from '../contexts/AuthContext';
 import ImageTitle from '../components/misc/ImageTitle';
-import PerformancesByGenre from '../components/performances/PerformancesByGenre';
-import PerformancesByTheaters2 from '../components/performances/PerformancesByTheaters2';
-import PerformancesNextWeek from '../components/performances/PerformancesNextWeek';
 import performancesService from '../services/performances.service';
+import PerformanceBrowse from '../components/performances/PerformanceBrowse';
 
 export default function BrowsingPage() {
-  const [performances, setPerformances] = useState([]);
+  const { user } = useContext(AuthContext);
   const [genres, setGenres] = useState([]);
-  const [searchParams] = useSearchParams();
   const rendered = useRef(false); // stops unnecessary rerender of performances state
-
-  const getPerformances = async () => {
-    const data = await performancesService.listAll();
-    setPerformances(data);
-  };
 
   const getGenres = async () => {
     const genreData = await performancesService.listAllGenre();
     setGenres(genreData);
   };
 
+  // Convert userBirthDate to a Date object and extract only the date part
+  const birthDate = new Date(user.BirthDate);
+  const today = new Date();
+  const tomorrow = new Date(today);
+  tomorrow.setDate(tomorrow.getDate() + 1);
+
+  const isSameDay = (date1, date2) =>
+    date1.getUTCDate() === date2.getUTCDate() && date1.getUTCMonth() === date2.getUTCMonth();
+
+  // Check if today is the user's birthday (ignoring the year)
+  const isBirthdayToday = isSameDay(birthDate, today);
+
   useEffect(() => {
     if (!rendered.current) {
       rendered.current = true;
       localStorage.setItem('empty_performance_img', '../../../public/Theatron.jpg');
-      getPerformances();
       getGenres();
     }
   }, []);
 
-  useEffect(() => {
-    getPerformances();
-  }, [searchParams]);
-
-  // Műfajok szétválasztása 3-nál több előadással rendelkezők és egyéb kategóriába
+  // Separation of genres into those with more than 3 performances and other categories
   const frequentGenres = genres.filter((genre) => genre.count >= 3);
   const rareGenres = genres.filter((genre) => genre.count < 3);
-
-  // Az "egyéb műfajok" előadásainak összegyűjtése
-  const otherPerformances = performances.filter((p) =>
-    p.genre.some((g) => rareGenres.some((rg) => rg.name === g.name))
-  );
 
   return (
     <>
@@ -53,30 +47,25 @@ export default function BrowsingPage() {
         description="Üdvözlünk a világ első színházi bérlet applikációjában! Válaszd ki a darabot, a helyed és már mehetsz is!"
       />
 
-      <div className="w-full max-w-screen-desktop flex flex-col items-center mx-auto">
-        {performances.length > 0 && <PerformancesNextWeek performances={performances} />}
-      </div>
-      <div className="w-full max-w-screen-desktop flex flex-col items-center mx-auto">
-        {performances.length > 0 && <PerformancesByTheaters2 performances={performances} />}
+      {/* Performances for the birth of the logged-in user */}
+      <div className="w-full max-w-screen-desktop flex flex-col items-center mx-auto ">
+        {isBirthdayToday ? <PerformanceBrowse params={{ startDate: today, endDate: tomorrow }} /> : null}
       </div>
 
-      {/* Legalább 3 előadásos műfajok */}
+      {/* Genres with at least 3 performances */}
       {frequentGenres.map((genre) => (
         <div
           key={genre.name}
           className="w-full max-w-screen-desktop flex flex-col items-center mx-auto"
         >
-          <PerformancesByGenre
-            performances={performances.filter((p) => p.genre.some((g) => g.name === genre.name))}
-            genre={genre}
-          />
+          <PerformanceBrowse params={{ genres: [genre.name] }} />
         </div>
       ))}
 
-      {/* Egyéb műfajok */}
-      {otherPerformances.length > 0 && (
+      {/* Other genres */}
+      {rareGenres.length > 0 && (
         <div className="w-full max-w-screen-desktop flex flex-col items-center mx-auto">
-          <PerformancesByGenre performances={otherPerformances} genre={{ name: 'Egyéb műfajok' }} />
+          <PerformanceBrowse params={{ genres: rareGenres.map((genre) => genre.name) }} />
         </div>
       )}
     </>
