+++ frontend/src/services/performances.service.js
@@ -36,9 +36,19 @@ const listAllGenre = async () => {
   }
 };
 
+const createPerformance = async (performanceData) => {
+  try {
+    const response = await axiosInstance.post('/api/performances', performanceData);
+    return response.data; // Successful answer
+  } catch (error) {
+    throw error.response ? error.response.data : error; // In case of error
+  }
+};
+
 export default {
   getById,
   list,
   listAll,
-  listAllGenre
+  listAllGenre,
+  createPerformance
 };
